refactor(cairo-parser): separate cli into own class
